chore: Move files for Core
