+++ Sources/PrimerSDK/Classes/Core/Analytics/AnalyticsEvent.swift
@@ -340,468 +340,6 @@ extension Analytics.Event {
 
 protocol AnalyticsEventProperties: Codable {}
 
-struct RawEventProperties: AnalyticsEventProperties {
-    private let values: [String: AnyCodable]
-
-    init(data: Data) throws {
-        values = try JSONDecoder().decode([String: AnyCodable].self, from: data)
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        values = try container.decode([String: AnyCodable].self)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        try container.encode(values)
-    }
-}
-
-struct MessageEventProperties: AnalyticsEventProperties {
-
-    var message: String?
-    var messageType: Analytics.Event.Property.MessageType
-    var severity: Analytics.Event.Property.Severity
-    var diagnosticsId: String?
-    var context: [String: Any]?
-
-    private enum CodingKeys: String, CodingKey {
-        case message
-        case messageType
-        case severity
-        case diagnosticsId
-        case context
-    }
-
-    fileprivate init(
-        message: String?,
-        messageType: Analytics.Event.Property.MessageType,
-        severity: Analytics.Event.Property.Severity,
-        diagnosticsId: String? = nil,
-        context: [String: Any]? = nil
-    ) {
-        self.message = message
-        self.messageType = messageType
-        self.severity = severity
-        self.diagnosticsId = diagnosticsId
-        self.context = context
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.message = try container.decodeIfPresent(String.self, forKey: .message)
-        self.messageType = try container.decode(Analytics.Event.Property.MessageType.self, forKey: .messageType)
-        self.severity = try container.decode(Analytics.Event.Property.Severity.self, forKey: .severity)
-        self.diagnosticsId = try container.decodeIfPresent(String.self, forKey: .diagnosticsId)
-        self.context = try container.decodeIfPresent([String: Any].self, forKey: .context)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encodeIfPresent(message, forKey: .message)
-        try container.encode(messageType, forKey: .messageType)
-        try container.encode(severity, forKey: .severity)
-        try container.encodeIfPresent(diagnosticsId, forKey: .diagnosticsId)
-        try container.encodeIfPresent(context, forKey: .context)
-    }
-}
-
-struct NetworkCallEventProperties: AnalyticsEventProperties {
-
-    var callType: Analytics.Event.Property.NetworkCallType
-    var id: String
-    var url: String
-    var method: HTTPMethod
-    var errorBody: String?
-    var responseCode: Int?
-    var params: [String: AnyCodable]?
-    var duration: TimeInterval?
-
-    private enum CodingKeys: String, CodingKey {
-        case callType
-        case id
-        case url
-        case method
-        case errorBody
-        case responseCode
-        case params
-        case duration
-    }
-
-    fileprivate init(
-        callType: Analytics.Event.Property.NetworkCallType,
-        id: String,
-        url: String,
-        method: HTTPMethod,
-        errorBody: String?,
-        responseCode: Int?,
-        duration: TimeInterval? = nil
-    ) {
-        self.callType = callType
-        self.id = id
-        self.url = url
-        self.method = method
-        self.errorBody = errorBody
-        self.responseCode = responseCode
-        self.duration = duration
-
-        let sdkProperties = SDKProperties()
-        if let sdkPropertiesDict = try? sdkProperties.asDictionary(),
-           let data = try? JSONSerialization.data(withJSONObject: sdkPropertiesDict, options: .fragmentsAllowed) {
-            let decoder = JSONDecoder()
-            if let anyDecodableDictionary = try? decoder.decode([String: AnyCodable].self, from: data) {
-                self.params = anyDecodableDictionary
-            }
-        } else {
-            self.params = nil
-        }
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.callType = try container.decode(Analytics.Event.Property.NetworkCallType.self, forKey: .callType)
-        self.id = try container.decode(String.self, forKey: .id)
-        self.url = try container.decode(String.self, forKey: .url)
-        self.method = try container.decode(HTTPMethod.self, forKey: .method)
-        self.errorBody = try container.decodeIfPresent(String.self, forKey: .errorBody)
-        self.responseCode = try container.decodeIfPresent(Int.self, forKey: .responseCode)
-        self.params = try container.decodeIfPresent([String: AnyCodable].self, forKey: .params)
-        self.duration = try container.decodeIfPresent(TimeInterval.self, forKey: .duration)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(callType, forKey: .callType)
-        try container.encode(id, forKey: .id)
-        try container.encode(url, forKey: .url)
-        try container.encode(method, forKey: .method)
-        try container.encodeIfPresent(errorBody, forKey: .errorBody)
-        try container.encodeIfPresent(responseCode, forKey: .responseCode)
-        try container.encodeIfPresent(params, forKey: .params)
-        try container.encodeIfPresent(duration, forKey: .duration)
-    }
-}
-
-struct NetworkConnectivityEventProperties: AnalyticsEventProperties {
-
-    var networkType: Connectivity.NetworkType
-    var params: [String: AnyCodable]?
-
-    private enum CodingKeys: String, CodingKey {
-        case networkType
-        case params
-    }
-
-    fileprivate init(networkType: Connectivity.NetworkType) {
-        self.networkType = networkType
-
-        let sdkProperties = SDKProperties()
-        if let sdkPropertiesDict = try? sdkProperties.asDictionary(),
-           let data = try? JSONSerialization.data(withJSONObject: sdkPropertiesDict, options: .fragmentsAllowed) {
-            let decoder = JSONDecoder()
-            if let anyDecodableDictionary = try? decoder.decode([String: AnyCodable].self, from: data) {
-                self.params = anyDecodableDictionary
-            }
-        } else {
-            self.params = nil
-        }
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.networkType = try container.decode(Connectivity.NetworkType.self, forKey: .networkType)
-        self.params = try container.decodeIfPresent([String: AnyCodable].self, forKey: .params)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(networkType, forKey: .networkType)
-        try container.encodeIfPresent(params, forKey: .params)
-    }
-}
-
-struct AppLifecycleEventProperties: AnalyticsEventProperties {
-
-    enum LifecycleType: String, Codable {
-        case backgrounded = "APP_DID_ENTER_BACKGROUND"
-        case foregrounded = "APP_WILL_ENTER_FOREGROUND"
-    }
-
-    let lifecycleType: LifecycleType
-    var params: [String: AnyCodable]?
-
-    fileprivate init(lifecycleType: LifecycleType) {
-        self.lifecycleType = lifecycleType
-        let sdkProperties = SDKProperties()
-        let dict = try? sdkProperties.asDictionary()
-        let data = try? JSONSerialization.data(withJSONObject: dict as Any, options: .fragmentsAllowed)
-        data.map { params = try? JSONDecoder().decode([String: AnyCodable].self, from: $0) }
-    }
-    
-}
-
-struct SDKEventProperties: AnalyticsEventProperties {
-
-    var name: String
-    var params: [String: AnyCodable]?
-
-    private enum CodingKeys: String, CodingKey {
-        case name
-        case params
-    }
-
-    fileprivate init(name: String, params: [String: String]?) {
-        self.name = name
-
-        var parameters: [String: Any] = params ?? [:]
-
-        let sdkProperties = SDKProperties()
-        if let sdkPropertiesDict = try? sdkProperties.asDictionary() {
-            parameters.merge(sdkPropertiesDict) {(current, _) in current}
-        }
-
-        if !parameters.isEmpty,
-           let parametersData = try? JSONSerialization.data(withJSONObject: parameters, options: .fragmentsAllowed) {
-            let decoder = JSONDecoder()
-            if let anyDecodableDictionary = try? decoder.decode([String: AnyCodable].self, from: parametersData) {
-                self.params = anyDecodableDictionary
-            }
-        } else {
-            self.params = nil
-        }
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.name = try container.decode(String.self, forKey: .name)
-        self.params = try container.decodeIfPresent([String: AnyCodable].self, forKey: .params)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(name, forKey: .name)
-        try container.encodeIfPresent(params, forKey: .params)
-    }
-}
-
-struct TimerEventProperties: AnalyticsEventProperties {
-
-    var momentType: Analytics.Event.Property.TimerType
-    var id: String?
-    var params: [String: AnyCodable]?
-    var duration: TimeInterval?
-    var context: [String: Any]?
-
-    private enum CodingKeys: String, CodingKey {
-        case momentType
-        case id
-        case params
-        case duration
-        case context
-    }
-
-    fileprivate init(
-        momentType: Analytics.Event.Property.TimerType,
-        id: String?,
-        duration: TimeInterval? = nil,
-        context: [String: Any]? = nil
-    ) {
-        self.momentType = momentType
-        self.id = id
-        self.duration = duration
-        self.context = context
-
-        let sdkProperties = SDKProperties()
-        if let sdkPropertiesDict = try? sdkProperties.asDictionary(),
-           let data = try? JSONSerialization.data(withJSONObject: sdkPropertiesDict, options: .fragmentsAllowed) {
-            let decoder = JSONDecoder()
-            if let anyDecodableDictionary = try? decoder.decode([String: AnyCodable].self, from: data) {
-                self.params = anyDecodableDictionary
-            }
-        } else {
-            self.params = nil
-        }
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.momentType = try container.decode(Analytics.Event.Property.TimerType.self, forKey: .momentType)
-        self.id = try container.decodeIfPresent(String.self, forKey: .id)
-        self.params = try container.decodeIfPresent([String: AnyCodable].self, forKey: .params)
-        self.duration = try container.decodeIfPresent(TimeInterval.self, forKey: .duration)
-        self.context = try container.decodeIfPresent([String: Any].self, forKey: .context)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(momentType, forKey: .momentType)
-        try container.encodeIfPresent(id, forKey: .id)
-        try container.encodeIfPresent(params, forKey: .params)
-        try container.encodeIfPresent(duration, forKey: .duration)
-        try container.encodeIfPresent(context, forKey: .context)
-    }
-}
-
-struct UIEventProperties: AnalyticsEventProperties {
-
-    var action: Analytics.Event.Property.Action
-    var context: Analytics.Event.Property.Context?
-    var extra: String?
-    var objectType: Analytics.Event.Property.ObjectType
-    var objectId: Analytics.Event.Property.ObjectId?
-    var objectClass: String?
-    var place: Analytics.Event.Property.Place
-    var params: [String: String]?
-
-    private enum CodingKeys: String, CodingKey {
-        case action
-        case context
-        case extra
-        case objectType
-        case objectId
-        case objectClass
-        case place
-        case params
-    }
-
-    fileprivate init(
-        action: Analytics.Event.Property.Action,
-        context: Analytics.Event.Property.Context?,
-        extra: String?,
-        objectType: Analytics.Event.Property.ObjectType,
-        objectId: Analytics.Event.Property.ObjectId?,
-        objectClass: String?,
-        place: Analytics.Event.Property.Place
-    ) {
-        self.action = action
-        self.context = context
-        self.extra = extra
-        self.objectType = objectType
-        self.objectId = objectId
-        self.objectClass = objectClass
-        self.place = place
-
-        if let jsonData = try? JSONEncoder().encode(SDKProperties()),
-           let jsonObject = try? JSONSerialization.jsonObject(with: jsonData, options: .allowFragments),
-           let params = jsonObject as? [String: String] {
-            self.params = params
-        }
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.action = try container.decode(Analytics.Event.Property.Action.self, forKey: .action)
-        self.context = try container.decodeIfPresent(Analytics.Event.Property.Context.self, forKey: .context)
-        self.extra = try container.decodeIfPresent(String.self, forKey: .extra)
-        self.objectType = try container.decode(Analytics.Event.Property.ObjectType.self, forKey: .objectType)
-        self.objectId = try container.decodeIfPresent(Analytics.Event.Property.ObjectId.self, forKey: .objectId)
-        self.objectClass = try container.decodeIfPresent(String.self, forKey: .objectClass)
-        self.place = try container.decode(Analytics.Event.Property.Place.self, forKey: .place)
-        self.params = try container.decodeIfPresent([String: String].self, forKey: .params)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(action, forKey: .action)
-        try container.encodeIfPresent(context, forKey: .context)
-        try container.encodeIfPresent(extra, forKey: .extra)
-        try container.encode(objectType, forKey: .objectType)
-        try container.encodeIfPresent(objectId, forKey: .objectId)
-        try container.encodeIfPresent(objectClass, forKey: .objectClass)
-        try container.encode(place, forKey: .place)
-        try container.encode(params, forKey: .params)
-    }
-}
-
-struct SDKProperties: Codable {
-
-    let clientToken: String?
-    let integrationType: String?
-    let paymentMethodType: String?
-    let sdkIntegrationType: PrimerSDKIntegrationType?
-    let sdkIntent: PrimerSessionIntent?
-    let sdkPaymentHandling: PrimerPaymentHandling?
-    let sdkSessionId: String?
-    let sdkSettings: [String: AnyCodable]?
-    let sdkType: String?
-    let sdkVersion: String?
-    let context: [String: AnyCodable]?
-
-    private enum CodingKeys: String, CodingKey {
-        case clientToken
-        case integrationType
-        case paymentMethodType
-        case sdkIntegrationType
-        case sdkIntent
-        case sdkPaymentHandling
-        case sdkSessionId
-        case sdkSettings
-        case sdkType
-        case sdkVersion
-        case context
-    }
-
-    fileprivate init() {
-        self.clientToken = AppState.current.clientToken
-        self.sdkIntegrationType = PrimerInternal.shared.sdkIntegrationType
-        #if COCOAPODS
-            self.integrationType = "COCOAPODS"
-        #else
-            self.integrationType = "SPM"
-        #endif
-        self.paymentMethodType = PrimerInternal.shared.selectedPaymentMethodType
-        self.sdkIntent = PrimerInternal.shared.intent
-        self.sdkPaymentHandling = PrimerSettings.current.paymentHandling
-        self.sdkSessionId = PrimerInternal.shared.checkoutSessionId
-
-        self.sdkType = Primer.shared.integrationOptions?.reactNativeVersion == nil ? "IOS_NATIVE" : "RN_IOS"
-        self.sdkVersion = VersionUtils.releaseVersionNumber
-        self.context = nil
-
-        if let settingsData = try? JSONEncoder().encode(PrimerSettings.current) {
-            let decoder = JSONDecoder()
-            if let anyDecodableDictionary = try? decoder.decode([String: AnyCodable].self, from: settingsData) {
-                self.sdkSettings = anyDecodableDictionary
-                return
-            }
-        }
-
-        self.sdkSettings = nil
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.clientToken = try container.decodeIfPresent(String.self, forKey: .clientToken)
-        self.integrationType = try container.decodeIfPresent(String.self, forKey: .integrationType)
-        self.paymentMethodType = try container.decodeIfPresent(String.self, forKey: .paymentMethodType)
-        self.sdkIntegrationType = try container.decodeIfPresent(PrimerSDKIntegrationType.self, forKey: .sdkIntegrationType)
-        self.sdkIntent = try container.decodeIfPresent(PrimerSessionIntent.self, forKey: .sdkIntent)
-        self.sdkPaymentHandling = try container.decodeIfPresent(PrimerPaymentHandling.self, forKey: .sdkPaymentHandling)
-        self.sdkSessionId = try container.decodeIfPresent(String.self, forKey: .sdkSessionId)
-        self.sdkSettings = try container.decodeIfPresent([String: AnyCodable].self, forKey: .sdkSettings)
-        self.sdkType = try container.decodeIfPresent(String.self, forKey: .sdkType)
-        self.sdkVersion = try container.decodeIfPresent(String.self, forKey: .sdkVersion)
-        self.context = try container.decodeIfPresent([String: AnyCodable].self, forKey: .context)
-
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encodeIfPresent(clientToken, forKey: .clientToken)
-        try container.encodeIfPresent(integrationType, forKey: .integrationType)
-        try container.encodeIfPresent(paymentMethodType, forKey: .paymentMethodType)
-        try container.encodeIfPresent(sdkIntent, forKey: .sdkIntent)
-        try container.encodeIfPresent(sdkPaymentHandling, forKey: .sdkPaymentHandling)
-        try container.encodeIfPresent(sdkSettings, forKey: .sdkSettings)
-        try container.encodeIfPresent(sdkSessionId, forKey: .sdkSessionId)
-        try container.encodeIfPresent(sdkSettings, forKey: .sdkSettings)
-        try container.encodeIfPresent(sdkType, forKey: .sdkType)
-        try container.encodeIfPresent(sdkVersion, forKey: .sdkVersion)
-        try container.encodeIfPresent(context, forKey: .context)
-    }
-}
-
 extension Analytics.Event {
 
     static func sdk(name: String, params: [String: String]?) -> Self {
