+++ Sources/PrimerSDK/Classes/Core/Analytics/RawEventProperties.swift
@@ -1,0 +1,26 @@
+//
+//  RawEventProperties.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+import Foundation
+@_spi(PrimerInternal) import PrimerFoundation
+
+struct RawEventProperties: AnalyticsEventProperties {
+    private let values: [String: AnyCodable]
+
+    init(data: Data) throws {
+        values = try JSONDecoder().decode([String: AnyCodable].self, from: data)
+    }
+
+    init(from decoder: Decoder) throws {
+        let container = try decoder.singleValueContainer()
+        values = try container.decode([String: AnyCodable].self)
+    }
+
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        try container.encode(values)
+    }
+}
