+++ Sources/PrimerSDK/Classes/Data Models/DismissalMechanism.swift
@@ -1,0 +1,9 @@
+//
+//  DismissalMechanism.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+public enum DismissalMechanism: Codable {
+    case gestures, closeButton
+}
