+++ Sources/PrimerSDK/Classes/Data Models/PrimerApiVersion.swift
@@ -1,0 +1,14 @@
+//
+//  PrimerApiVersion.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+// MARK: - API Version
+// swiftlint:disable identifier_name
+public enum PrimerApiVersion: String, Codable {
+    case V2_4 = "2.4"
+
+    public static let latest = PrimerApiVersion.V2_4
+}
+// swiftlint:enable identifier_name
