+++ Sources/PrimerSDK/Classes/Data Models/PrimerAppearanceMode.swift
@@ -1,0 +1,11 @@
+//
+//  PrimerAppearanceMode.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+public enum PrimerAppearanceMode: String, Codable {
+    case system = "SYSTEM"
+    case light = "LIGHT"
+    case dark = "DARK"
+}
