+++ Sources/PrimerSDK/Classes/Data Models/PrimerPaymentHandling.swift
@@ -1,0 +1,10 @@
+//
+//  PrimerPaymentHandling.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+public enum PrimerPaymentHandling: String, Codable {
+    case auto   = "AUTO"
+    case manual = "MANUAL"
+}
