+++ Sources/PrimerSDK/Classes/Data Models/PrimerThreeDsOptions.swift
@@ -1,0 +1,13 @@
+//
+//  PrimerThreeDsOptions.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+public struct PrimerThreeDsOptions: Codable {
+    let threeDsAppRequestorUrl: String?
+
+    public init(threeDsAppRequestorUrl: String? = nil) {
+        self.threeDsAppRequestorUrl = threeDsAppRequestorUrl
+    }
+}
