+++ Sources/PrimerSDK/Classes/Error Handler/Primer3DSErrorInfo.swift
@@ -1,0 +1,17 @@
+//
+//  Primer3DSErrorInfo.swift
+//
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
+//  Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+public struct Primer3DSErrorInfo {
+    let errorId: String
+    let errorDescription: String
+    let recoverySuggestion: String?
+    let threeDsErrorCode: Int?
+    let threeDsErrorType: String?
+    let threeDsErrorComponent: String?
+    let threeDsSdkTranscationId: String?
+    let threeDsSErrorVersion: String?
+    let threeDsErrorDetail: String?
+}
